Revert "Van nnt login api done"
