+++ CM.sql
@@ -1,7 +1,7 @@
 create database if not exists cm;
 use cm;
 create table accounts (
-    user_id bigint(20) not null primary key auto_increment,
+    user_id bigint(20) not null primary key,
     username varchar(50),
     password varchar(50),
     name_of_customer varchar(128),
@@ -33,22 +33,9 @@ create table schedules (
 );
 
 create table seat (
-	seat_id bigint(20) not null primary key auto_increment,
+	seat_id bigint(20) not null primary key,
     seat_name varchar(4),
     seat_status boolean,
     schedule_id bigint(20),
     constraint fk_schedule foreign key (schedule_id) references schedules(schedule_id)
-);
-
-create table ticket (
-   ticket_id bigint(20) not null primary key auto_increment,
-   username varchar(50),
-   phone varchar(50),
-   room integer(2),
-   start_date date,
-   start_time time,
-   theatre varchar(50),
-   movie_name varchar(200),
-   amount integer(2),
-   total_price varchar(50)
 );
