--- src/main/java/com/cm/SpringBootRestApiApp.java
@@ -1,21 +1,0 @@
-package com.cm;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-import com.cm.config.JpaConfiguration;
-
-@Import(JpaConfiguration.class)
-@SpringBootApplication(scanBasePackages = {"com.cm" })	
-public class SpringBootRestApiApp {
-	
-	public static void main(String[] args) {
-		SpringApplication.run(SpringBootRestApiApp.class, args);
-	}
-
-}
