--- src/main/java/com/cm/common/CustomeErrType.java
@@ -1,14 +1,0 @@
-package com.cm.common;
-
-public class CustomeErrType {
-
-	private String errorMessage;
-
-    public void CustomErrType(String errorMessage){
-        this.errorMessage = errorMessage;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-}
