--- src/main/java/com/cm/constants/ParamConstants.java
@@ -1,7 +1,0 @@
-package com.cm.constants;
-
-public class ParamConstants {
-	//Account
-	public static final String USERNAME = "username";
-	public static final String PASSWORD = "password";
-}
