--- src/main/java/com/cm/constants/URLConstants.java
@@ -1,6 +1,0 @@
-package com.cm.constants;
-
-public class URLConstants {
-	public static final String HOME_URL = "/home";
-	public static final String AUTHENTICATION_URL = "/authentication";
-}
