--- src/main/java/com/cm/entity/TicketEntity.java
@@ -1,143 +1,0 @@
-package com.cm.entity;
-// Generated Feb 8, 2017 3:09:26 PM by Hibernate Tools 4.0.0.Final
-
-import java.util.Date;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import static javax.persistence.GenerationType.IDENTITY;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-/**
- * Ticket generated by hbm2java
- */
-@Entity
-@Table(name = "ticket", catalog = "cm")
-public class TicketEntity implements java.io.Serializable {
-
-	private Long ticketId;
-	private String username;
-	private String phone;
-	private Integer room;
-	private Date startDate;
-	private Date startTime;
-	private String theatre;
-	private String movieName;
-	private Integer amount;
-	private String totalPrice;
-
-	public TicketEntity() {
-	}
-
-	public TicketEntity(String username, String phone, Integer room, Date startDate, Date startTime, String theatre,
-			String movieName, Integer amount, String totalPrice) {
-		this.username = username;
-		this.phone = phone;
-		this.room = room;
-		this.startDate = startDate;
-		this.startTime = startTime;
-		this.theatre = theatre;
-		this.movieName = movieName;
-		this.amount = amount;
-		this.totalPrice = totalPrice;
-	}
-
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-
-	@Column(name = "ticket_id", unique = true, nullable = false)
-	public Long getTicketId() {
-		return this.ticketId;
-	}
-
-	public void setTicketId(Long ticketId) {
-		this.ticketId = ticketId;
-	}
-
-	@Column(name = "username", length = 50)
-	public String getUsername() {
-		return this.username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	@Column(name = "phone", length = 50)
-	public String getPhone() {
-		return this.phone;
-	}
-
-	public void setPhone(String phone) {
-		this.phone = phone;
-	}
-
-	@Column(name = "room")
-	public Integer getRoom() {
-		return this.room;
-	}
-
-	public void setRoom(Integer room) {
-		this.room = room;
-	}
-
-	@Temporal(TemporalType.DATE)
-	@Column(name = "start_date", length = 10)
-	public Date getStartDate() {
-		return this.startDate;
-	}
-
-	public void setStartDate(Date startDate) {
-		this.startDate = startDate;
-	}
-
-	@Temporal(TemporalType.TIME)
-	@Column(name = "start_time", length = 8)
-	public Date getStartTime() {
-		return this.startTime;
-	}
-
-	public void setStartTime(Date startTime) {
-		this.startTime = startTime;
-	}
-
-	@Column(name = "theatre", length = 50)
-	public String getTheatre() {
-		return this.theatre;
-	}
-
-	public void setTheatre(String theatre) {
-		this.theatre = theatre;
-	}
-
-	@Column(name = "movie_name", length = 200)
-	public String getMovieName() {
-		return this.movieName;
-	}
-
-	public void setMovieName(String movieName) {
-		this.movieName = movieName;
-	}
-
-	@Column(name = "amount")
-	public Integer getAmount() {
-		return this.amount;
-	}
-
-	public void setAmount(Integer amount) {
-		this.amount = amount;
-	}
-
-	@Column(name = "total_price", length = 50)
-	public String getTotalPrice() {
-		return this.totalPrice;
-	}
-
-	public void setTotalPrice(String totalPrice) {
-		this.totalPrice = totalPrice;
-	}
-
-}
