--- src/main/java/com/cm/model/interfaces/AccountModelInterface.java
@@ -1,7 +1,0 @@
-package com.cm.model.interfaces;
-
-import com.cm.entity.AccountEntity;
-
-public interface AccountModelInterface {
-	AccountEntity getAccountByUsername(String username);
-}
