+++ src/main/java/com/fpt/cm/App.java
@@ -1,0 +1,13 @@
+package com.fpt.cm;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
