+++ src/main/java/com/fpt/cm/common/constant/ParamConstant.java
@@ -1,0 +1,7 @@
+package com.fpt.cm.common.constant;
+
+public class ParamConstant {
+	//Account Info
+	public static final String USERNAME = "username";
+	public static final String PASSWORD = "password";
+}
