+++ src/main/java/com/fpt/cm/common/constant/URLConstant.java
@@ -1,0 +1,5 @@
+package com.fpt.cm.common.constant;
+
+public class URLConstant {
+	public static final String AUTHENTICATION = "/login";
+}
