+++ src/main/java/com/fpt/cm/entities/AccountEntity.java
@@ -1,0 +1,115 @@
+package com.fpt.cm.entities;
+// Generated Feb 7, 2017 7:37:11 PM by Hibernate Tools 4.3.1.Final
+
+import java.util.Date;
+import java.util.HashSet;
+import java.util.Set;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+
+/**
+ * Accounts generated by hbm2java
+ */
+@Entity
+@Table(name = "accounts", catalog = "cm")
+public class AccountEntity implements java.io.Serializable {
+
+	private long userId;
+	private String username;
+	private String password;
+	private String nameOfCustomer;
+	private Date birthdate;
+	private String role;
+	private Set<MovieEntity> movies = new HashSet<MovieEntity>(0);
+
+	public AccountEntity() {
+	}
+
+	public AccountEntity(long userId) {
+		this.userId = userId;
+	}
+
+	public AccountEntity(long userId, String username, String password, String nameOfCustomer, Date birthdate, String role,
+			Set<MovieEntity> movies) {
+		this.userId = userId;
+		this.username = username;
+		this.password = password;
+		this.nameOfCustomer = nameOfCustomer;
+		this.birthdate = birthdate;
+		this.role = role;
+		this.movies = movies;
+	}
+
+	@Id
+
+	@Column(name = "user_id", unique = true, nullable = false)
+	public long getUserId() {
+		return this.userId;
+	}
+
+	public void setUserId(long userId) {
+		this.userId = userId;
+	}
+
+	@Column(name = "username", length = 50)
+	public String getUsername() {
+		return this.username;
+	}
+
+	public void setUsername(String username) {
+		this.username = username;
+	}
+
+	@Column(name = "password", length = 50)
+	public String getPassword() {
+		return this.password;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+	@Column(name = "name_of_customer", length = 128)
+	public String getNameOfCustomer() {
+		return this.nameOfCustomer;
+	}
+
+	public void setNameOfCustomer(String nameOfCustomer) {
+		this.nameOfCustomer = nameOfCustomer;
+	}
+
+	@Temporal(TemporalType.DATE)
+	@Column(name = "birthdate", length = 10)
+	public Date getBirthdate() {
+		return this.birthdate;
+	}
+
+	public void setBirthdate(Date birthdate) {
+		this.birthdate = birthdate;
+	}
+
+	@Column(name = "role", length = 30)
+	public String getRole() {
+		return this.role;
+	}
+
+	public void setRole(String role) {
+		this.role = role;
+	}
+
+	@OneToMany(fetch = FetchType.LAZY, mappedBy = "accounts")
+	public Set<MovieEntity> getMovies() {
+		return this.movies;
+	}
+
+	public void setMovies(Set<MovieEntity> movies) {
+		this.movies = movies;
+	}
+
+}
