+++ src/main/java/com/fpt/cm/entities/MovieEntity.java
@@ -1,0 +1,161 @@
+package com.fpt.cm.entities;
+// Generated Feb 7, 2017 7:37:11 PM by Hibernate Tools 4.3.1.Final
+
+import java.util.Date;
+import java.util.HashSet;
+import java.util.Set;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import static javax.persistence.GenerationType.IDENTITY;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+
+/**
+ * Movie generated by hbm2java
+ */
+@Entity
+@Table(name = "movie", catalog = "cm")
+public class MovieEntity implements java.io.Serializable {
+
+	private Long movieId;
+	private AccountEntity accounts;
+	private String movieName;
+	private String introduction;
+	private String actor;
+	private String genre;
+	private Date startDate;
+	private Date endDate;
+	private String trailer;
+	private String picture;
+	private Set<ScheduleEntity> scheduleses = new HashSet<ScheduleEntity>(0);
+
+	public MovieEntity() {
+	}
+
+	public MovieEntity(AccountEntity accounts, String movieName, String introduction, String actor, String genre, Date startDate,
+			Date endDate, String trailer, String picture, Set<ScheduleEntity> scheduleses) {
+		this.accounts = accounts;
+		this.movieName = movieName;
+		this.introduction = introduction;
+		this.actor = actor;
+		this.genre = genre;
+		this.startDate = startDate;
+		this.endDate = endDate;
+		this.trailer = trailer;
+		this.picture = picture;
+		this.scheduleses = scheduleses;
+	}
+
+	@Id
+	@GeneratedValue(strategy = IDENTITY)
+
+	@Column(name = "movie_id", unique = true, nullable = false)
+	public Long getMovieId() {
+		return this.movieId;
+	}
+
+	public void setMovieId(Long movieId) {
+		this.movieId = movieId;
+	}
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "add_by_id")
+	public AccountEntity getAccounts() {
+		return this.accounts;
+	}
+
+	public void setAccounts(AccountEntity accounts) {
+		this.accounts = accounts;
+	}
+
+	@Column(name = "movie_name", length = 200)
+	public String getMovieName() {
+		return this.movieName;
+	}
+
+	public void setMovieName(String movieName) {
+		this.movieName = movieName;
+	}
+
+	@Column(name = "introduction", length = 2000)
+	public String getIntroduction() {
+		return this.introduction;
+	}
+
+	public void setIntroduction(String introduction) {
+		this.introduction = introduction;
+	}
+
+	@Column(name = "actor", length = 200)
+	public String getActor() {
+		return this.actor;
+	}
+
+	public void setActor(String actor) {
+		this.actor = actor;
+	}
+
+	@Column(name = "genre", length = 50)
+	public String getGenre() {
+		return this.genre;
+	}
+
+	public void setGenre(String genre) {
+		this.genre = genre;
+	}
+
+	@Temporal(TemporalType.DATE)
+	@Column(name = "start_date", length = 10)
+	public Date getStartDate() {
+		return this.startDate;
+	}
+
+	public void setStartDate(Date startDate) {
+		this.startDate = startDate;
+	}
+
+	@Temporal(TemporalType.DATE)
+	@Column(name = "end_date", length = 10)
+	public Date getEndDate() {
+		return this.endDate;
+	}
+
+	public void setEndDate(Date endDate) {
+		this.endDate = endDate;
+	}
+
+	@Column(name = "trailer", length = 300)
+	public String getTrailer() {
+		return this.trailer;
+	}
+
+	public void setTrailer(String trailer) {
+		this.trailer = trailer;
+	}
+
+	@Column(name = "picture", length = 300)
+	public String getPicture() {
+		return this.picture;
+	}
+
+	public void setPicture(String picture) {
+		this.picture = picture;
+	}
+
+	@OneToMany(fetch = FetchType.LAZY, mappedBy = "movie")
+	public Set<ScheduleEntity> getScheduleses() {
+		return this.scheduleses;
+	}
+
+	public void setScheduleses(Set<ScheduleEntity> scheduleses) {
+		this.scheduleses = scheduleses;
+	}
+
+}
