+++ src/main/java/com/fpt/cm/entities/SeatEntity.java
@@ -1,0 +1,77 @@
+package com.fpt.cm.entities;
+// Generated Feb 7, 2017 7:37:11 PM by Hibernate Tools 4.3.1.Final
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+/**
+ * Seat generated by hbm2java
+ */
+@Entity
+@Table(name = "seat", catalog = "cm")
+public class SeatEntity implements java.io.Serializable {
+
+	private long seatId;
+	private ScheduleEntity schedules;
+	private String seatName;
+	private Boolean seatStatus;
+
+	public SeatEntity() {
+	}
+
+	public SeatEntity(long seatId) {
+		this.seatId = seatId;
+	}
+
+	public SeatEntity(long seatId, ScheduleEntity schedules, String seatName, Boolean seatStatus) {
+		this.seatId = seatId;
+		this.schedules = schedules;
+		this.seatName = seatName;
+		this.seatStatus = seatStatus;
+	}
+
+	@Id
+
+	@Column(name = "seat_id", unique = true, nullable = false)
+	public long getSeatId() {
+		return this.seatId;
+	}
+
+	public void setSeatId(long seatId) {
+		this.seatId = seatId;
+	}
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "schedule_id")
+	public ScheduleEntity getSchedules() {
+		return this.schedules;
+	}
+
+	public void setSchedules(ScheduleEntity schedules) {
+		this.schedules = schedules;
+	}
+
+	@Column(name = "seat_name", length = 4)
+	public String getSeatName() {
+		return this.seatName;
+	}
+
+	public void setSeatName(String seatName) {
+		this.seatName = seatName;
+	}
+
+	@Column(name = "seat_status")
+	public Boolean getSeatStatus() {
+		return this.seatStatus;
+	}
+
+	public void setSeatStatus(Boolean seatStatus) {
+		this.seatStatus = seatStatus;
+	}
+
+}
