+++ src/main/java/com/fpt/cm/model/common/commonDAO.java
@@ -1,0 +1,14 @@
+package com.fpt.cm.model.common;
+
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+
+public abstract class commonDAO {
+	@Autowired
+	private SessionFactory sessionFactory;
+	
+	protected Session getSession() {
+		return sessionFactory.getCurrentSession();
+	}
+}
