+++ src/main/java/com/fpt/cm/model/interfaces/AccountModelInterface.java
@@ -1,0 +1,8 @@
+package com.fpt.cm.model.interfaces;
+
+import com.fpt.cm.entities.AccountEntity;
+
+public interface AccountModelInterface {
+	AccountEntity getAccountByUsername(String username);
+	
+}
