+++ src/main/java/com/fpt/cm/service/interfaces/AccountServiceInterface.java
@@ -1,0 +1,8 @@
+package com.fpt.cm.service.interfaces;
+
+import com.fpt.cm.entities.AccountEntity;
+
+public interface AccountServiceInterface {
+	AccountEntity getAccountByUsername (String username);
+	AccountEntity authenticate(String username, String password);
+}
