+++ src/main/java/hibernate.cfg.xml
@@ -6,16 +6,15 @@
     <session-factory>
         <property name="hibernate.bytecode.use_reflection_optimizer">false</property>
         <property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
-        <property name="hibernate.connection.password">123456</property>
+        <property name="hibernate.connection.password">thdb010708</property>
         <property name="hibernate.connection.url">jdbc:mysql://localhost:3306/cm</property>
         <property name="hibernate.connection.username">root</property>
-        <property name="hibernate.dialect">org.hibernate.dialect.MySQL5Dialect</property>
+        <property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
         <property name="hibernate.search.autoregister_listeners">false</property>
         <property name="hibernate.validator.apply_to_ddl">false</property>
-        <mapping class="com.cm.entity.Seat" />
-        <mapping class="com.cm.entity.Accounts" />
-        <mapping class="com.cm.entity.Schedules" />
-        <mapping class="com.cm.entity.Movie" />
-        <mapping class="com.cm.entity.Ticket" />
+        <mapping class="com.fpt.entities.Seat" />
+        <mapping class="com.fpt.entities.Movie" />
+        <mapping class="com.fpt.entities.Schedules" />
+        <mapping class="com.fpt.entities.Accounts" />
     </session-factory>
 </hibernate-configuration>
