--- src/main/resources/application.yml
@@ -1,20 +1,0 @@
----
-server:
-  port: 8080
-  contextPath: /cm
----
-spring:
-  profiles: default
-datasource:
-  sampleapp:
-    url: jdbc:mysql://localhost:3306/cm
-    username: root
-    password: 123456
-    driverClassName: com.mysql.jdbc.Driver
-    defaultSchema:
-    maxPoolSize: 20
-    hibernate:
-      hbm2ddl.method: update
-      show_sql: true
-      format_sql: true
-      dialect: org.hibernate.dialect.MySQLDialect
